requirements.txt
aioredis==2.0.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
black==24.3.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
envparse==0.2.0
exceptiongroup==1.2.0
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
jwcrypto==1.5.6
Mako==1.3.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.0
prometheus_client==0.20.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
starlette_exporter==0.21.0
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
